+++ .env.sample
@@ -10,3 +10,12 @@ CONSUMER_KEY =
 CONSUMER_SECRET =
 ACCESS_TOKEN = 
 ACCESS_TOKEN_SECRET = 
+
+[OPTIONAL]
+START_MESSAGE =
+START_MEDIA =
+CUSTOM_TEXT =
+
+BUTTON_TITLE =
+TAKE_REPLIES =\
+TAKE_RETWEETS =
