+++ runtime.txt
@@ -1,1 +1,1 @@
-python-3.9.5
+python-3.9.6
